refactor(system): move Homebrew detection
